Refactor and add reset route
