+++ .golangci.yml
@@ -14,6 +14,7 @@ linters:
     - test
     - unused
   disable:
+    - cyclop
     - depguard
     - exhaustruct
     - gochecknoglobals
@@ -25,3 +26,10 @@ linters:
     - varnamelen
     - wrapcheck
     - wsl
+
+linters-settings:
+  tagliatelle:
+    case:
+      use-field-name: true
+      rules:
+        json: snake
